refactor(partners): move supplier boundary out of pms
