+++ app/partners/export/suppliers/contracts/__init__.py
@@ -1,0 +1,1 @@
+# app/partners/export/suppliers/contracts/__init__.py
