+++ app/partners/suppliers/__init__.py
@@ -1,0 +1,1 @@
+# app/partners/suppliers/__init__.py
