+++ app/partners/suppliers/routers/__init__.py
@@ -1,0 +1,1 @@
+# app/partners/suppliers/routers/__init__.py
