--- app/pms/public/suppliers/__init__.py
@@ -1,1 +1,0 @@
-# app/pms/public/suppliers/__init__.py
