--- app/pms/public/suppliers/contracts/__init__.py
@@ -1,1 +1,0 @@
-# app/pms/public/suppliers/contracts/__init__.py
