--- app/pms/public/suppliers/services/__init__.py
@@ -1,1 +1,0 @@
-# app/pms/public/suppliers/services/__init__.py
