--- app/pms/suppliers/__init__.py
@@ -1,1 +1,0 @@
-# app/pms/suppliers/__init__.py
