--- app/pms/suppliers/contracts/__init__.py
@@ -1,1 +1,0 @@
-# app/pms/suppliers/contracts/__init__.py
