--- app/pms/suppliers/models/__init__.py
@@ -1,8 +1,0 @@
-# app/pms/suppliers/models/__init__.py
-from app.pms.suppliers.models.supplier import Supplier
-from app.pms.suppliers.models.supplier_contact import SupplierContact
-
-__all__ = [
-    "Supplier",
-    "SupplierContact",
-]
