--- app/pms/suppliers/models/supplier_contact.py
@@ -1,40 +1,0 @@
-# app/pms/suppliers/models/supplier_contact.py
-from __future__ import annotations
-
-from datetime import datetime
-from typing import Optional
-
-from sqlalchemy import Boolean, DateTime, ForeignKey, Integer, String, func
-from sqlalchemy.orm import Mapped, mapped_column, relationship
-
-from app.db.base import Base
-
-
-class SupplierContact(Base):
-    __tablename__ = "supplier_contacts"
-
-    id: Mapped[int] = mapped_column(Integer, primary_key=True, autoincrement=True)
-
-    supplier_id: Mapped[int] = mapped_column(
-        Integer,
-        ForeignKey("suppliers.id", ondelete="RESTRICT"),
-        nullable=False,
-    )
-
-    name: Mapped[str] = mapped_column(String(100), nullable=False)
-    phone: Mapped[Optional[str]] = mapped_column(String(50), nullable=True)
-    email: Mapped[Optional[str]] = mapped_column(String(255), nullable=True)
-    wechat: Mapped[Optional[str]] = mapped_column(String(64), nullable=True)
-
-    role: Mapped[str] = mapped_column(String(32), nullable=False, default="other")
-    is_primary: Mapped[bool] = mapped_column(Boolean, nullable=False, default=False)
-    active: Mapped[bool] = mapped_column(Boolean, nullable=False, default=True)
-
-    supplier = relationship("Supplier", back_populates="contacts")
-
-    created_at: Mapped[datetime] = mapped_column(
-        DateTime(timezone=True), nullable=False, server_default=func.now()
-    )
-    updated_at: Mapped[datetime] = mapped_column(
-        DateTime(timezone=True), nullable=False, server_default=func.now(), onupdate=func.now()
-    )
