--- app/pms/suppliers/repos/__init__.py
@@ -1,1 +1,0 @@
-# app/pms/suppliers/repos/__init__.py
